feat(api): migrate get admin certification center memberships route
