+++ api/src/team/application/certification-center-membership/certification-center-membership.admin.route.js
@@ -82,6 +82,35 @@ export const certificationCenterMembershipAdminRoutes = [
     },
   },
   {
+    method: 'GET',
+    path: '/api/admin/certification-centers/{certificationCenterId}/certification-center-memberships',
+    config: {
+      pre: [
+        {
+          method: (request, h) =>
+            securityPreHandlers.hasAtLeastOneAccessOf([
+              securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
+              securityPreHandlers.checkAdminMemberHasRoleCertif,
+              securityPreHandlers.checkAdminMemberHasRoleSupport,
+              securityPreHandlers.checkAdminMemberHasRoleMetier,
+            ])(request, h),
+          assign: 'hasAuthorizationToAccessAdminScope',
+        },
+      ],
+      validate: {
+        params: Joi.object({
+          certificationCenterId: identifiersType.certificationCenterId,
+        }),
+      },
+      handler: certificationCenterMembershipAdminController.findCertificationCenterMembershipsByCertificationCenter,
+      notes: [
+        "- **Cette route est restreinte aux utilisateurs ayant les droits d'accès**\n" +
+          "- Récupération de tous les membres d'un centre de certification.\n",
+      ],
+      tags: ['api', 'admin', 'certification-center-membership'],
+    },
+  },
+  {
     method: 'POST',
     path: '/api/admin/certification-centers/{certificationCenterId}/certification-center-memberships',
     config: {
