+++ api/tests/unit/application/certification-centers/certification-center-controller_test.js
@@ -1,47 +1,8 @@
 import { certificationCenterController } from '../../../../lib/application/certification-centers/certification-center-controller.js';
 import { usecases } from '../../../../lib/domain/usecases/index.js';
-import { usecases as teamUsecases } from '../../../../src/team/domain/usecases/index.js';
 import { expect, hFake, sinon } from '../../../test-helper.js';
 
 describe('Unit | Controller | certifications-center-controller', function () {
-  describe('#findCertificationCenterMembershipsByCertificationCenter', function () {
-    const certificationCenterId = 1;
-
-    const request = {
-      params: {
-        certificationCenterId,
-      },
-    };
-    let certificationCenterMembershipSerializerStub;
-
-    beforeEach(function () {
-      sinon.stub(teamUsecases, 'findCertificationCenterMembershipsByCertificationCenter');
-      certificationCenterMembershipSerializerStub = {
-        serialize: sinon.stub(),
-      };
-    });
-
-    it('should call usecase and serializer and return ok', async function () {
-      // given
-      teamUsecases.findCertificationCenterMembershipsByCertificationCenter
-        .withArgs({
-          certificationCenterId,
-        })
-        .resolves({});
-      certificationCenterMembershipSerializerStub.serialize.withArgs({}).returns('ok');
-
-      // when
-      const response = await certificationCenterController.findCertificationCenterMembershipsByCertificationCenter(
-        request,
-        hFake,
-        { certificationCenterMembershipSerializer: certificationCenterMembershipSerializerStub },
-      );
-
-      // then
-      expect(response).to.equal('ok');
-    });
-  });
-
   describe('#updateReferer', function () {
     it('should call updateCertificationCenterReferer usecase and return 204', async function () {
       // given
